# docs/contracts/dre-narrative.openapi.yml
openapi: 3.0.3
info:
  title: DRE & Narrative — Aicfo API
  description: Contract for the dre-narrative module. Extracted from the canonical Aicfo OpenAPI spec.
  version: 0.2.0
servers:
  - url: http://localhost:3000
paths:
...
  /analysis/{analysisId}/status:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: analysisId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  referenceMonth:
                    type: string
                  status:
                    type: string
                  mode:
                    type: string
                  hasActionPlan:
                    type: boolean
                  actionItemCount:
                    type: number
                  generatedAt:
                    type: string
                    format: date-time
                    nullable: true
                  deliveredAt:
                    type: string
                    format: date-time
                    nullable: true
                required:
                  - id
                  - referenceMonth
                  - status
                  - mode
                  - hasActionPlan
                  - actionItemCount
                  - generatedAt
                  - deliveredAt
                additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '500':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
  /analysis/{analysisId}/retry:
    post:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: analysisId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                required:
                  - id
                  - status
                additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '500':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT